Organized all classes into distinct folders; fixed Gson import issue in ProfileFragment
