+++ app/src/main/java/com/example/circleapp/EventDisplay/YourEventsFragment.java
@@ -1,4 +1,4 @@
-package com.example.circleapp;
+package com.example.circleapp.EventDisplay;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,6 +10,10 @@ import android.widget.Button;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.circleapp.BaseObjects.Event;
+import com.example.circleapp.QRCode.ScanQRActivity;
+import com.example.circleapp.R;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,7 +31,7 @@ public class YourEventsFragment extends Fragment {
         Button scanButton = rootView.findViewById(R.id.scan_button);
 
         scanButton.setOnClickListener(v -> {
-            Intent intent = new Intent(rootView.getContext(), com.example.circleapp.ScanQRActivity.class);
+            Intent intent = new Intent(rootView.getContext(), ScanQRActivity.class);
             startActivity(intent);
         });
