+++ app/src/main/java/com/example/circleapp/Profile/MakeProfileActivity.java
@@ -1,4 +1,4 @@
-package com.example.circleapp;
+package com.example.circleapp.Profile;
 
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
@@ -15,6 +15,10 @@ import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
+import com.example.circleapp.BaseObjects.Attendee;
+import com.example.circleapp.Firebase.FirebaseManager;
+import com.example.circleapp.Firebase.PreferenceUtils;
+import com.example.circleapp.R;
 import com.github.dhaval2404.imagepicker.ImagePicker;
 
 // Activity that starts when User wants to make a profile
